Extract evalKql util from workflow to a shared platform plugin
