+++ package.json
@@ -611,6 +611,7 @@
     "@kbn/esql-ux-example-plugin": "link:examples/esql_ux_example",
     "@kbn/esql-validation-example-plugin": "link:examples/esql_validation_example",
     "@kbn/eui-provider-dev-warning": "link:src/platform/test/plugin_functional/plugins/eui_provider_dev_warning",
+    "@kbn/eval-kql": "link:src/platform/packages/shared/kbn-eval-kql",
     "@kbn/event-annotation-common": "link:src/platform/packages/shared/kbn-event-annotation-common",
     "@kbn/event-annotation-components": "link:src/platform/packages/shared/kbn-event-annotation-components",
     "@kbn/event-annotation-listing-plugin": "link:src/platform/plugins/private/event_annotation_listing",
