+++ src/platform/packages/shared/kbn-eval-kql/README.md
@@ -1,0 +1,3 @@
+# @kbn/eval-kql
+
+Empty package generated by @kbn/generate
