+++ src/platform/packages/shared/kbn-eval-kql/index.ts
@@ -7,9 +7,4 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { parseDuration } from './parse-duration/parse-duration';
-export { buildStepExecutionId } from './build_step_execution_id/build_step_execution_id';
-export { stringifyStackFrames } from './stringify_stack_frames';
-export { getKibanaUrl, buildWorkflowExecutionUrl } from './get_kibana_url';
-export { generateExecutionTaskScope } from './generate_execution_task_scope';
-export { TimeoutAbortedError, abortableTimeout } from './abortable_timeout/abortable_timeout';
+export { evaluateKql } from './src/eval_kql';
