+++ src/platform/packages/shared/kbn-eval-kql/kibana.jsonc
@@ -1,0 +1,7 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/eval-kql",
+  "owner": "@elastic/workflows-eng",
+  "group": "platform",
+  "visibility": "shared"
+}
