+++ src/platform/packages/shared/kbn-eval-kql/src/eval_kql.test.ts
