+++ src/platform/packages/shared/kbn-eval-kql/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "@kbn/tsconfig-base/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "target/types",
+  },
+  "include": [
+    "**/*.ts",
+  ],
+  "exclude": [
+    "target/**/*"
+  ],
+  "kbn_references": [
+    "@kbn/datemath",
+    "@kbn/es-query",
+    "@kbn/workflows",
+  ]
+}
